backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
python-dotenv>=1.0.0
google-generativeai>=0.8.0
big-O>=0.11.0
